Updated file names and reorganized folders
